[BGPT] Make render surface for opacity etc. simpler and more accurate
The new algorithm takes advantage of an attribute of the cc property
tree: a child always has a bigger id than its parent, which is simpler
and has higher performance.

It also fix a bug that sometimes we mistakenly ignored child render
surfaces as a controlled layer. Now always count child render surface
as a controlled layer of the parent.

Bug: 955177
Change-Id: If0b9d657ddfbf69dd4f099aa85f52427479c6fbc